requirements.txt
requests==2.22.0
beautifulsoup4==4.9.1
Pillow==7.1.2
